streaming/ray_stm/requirements.txt
# python=3.11.5
ray
redis
lru-dict
PyYaml
confluent-kafka